requirements.txt
# Backend Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0

# Base de Datos
mysql-connector-python==8.2.0
sqlalchemy==2.0.23

# Video y Visión Computacional
opencv-python==4.8.1.78
numpy==1.24.3

# Para futuro (YOLOv8 y procesamiento)
# ultralytics==8.0.0
# torch>=2.0.0
# numpy>=1.24.0
